Refactor #team_names
In general, you want to avoid writing your methods so that they depend
on any kind of state. In this case, I refactored your `#team_name`
method so that it has one input and returns one output.

It no longer relies on the @teams variable existing and it’s much more
clear where the @teams variable is getting its value from (the return
value of the team_names method).

When we get into testing, writing your code this way will also make it
much easier to test.