Refactor #stats and #team_names
We can combine the #stats and #team_names methods into one method since
that is the only place that you’re using the `@teams` array. It’s now
much more explicit how the value for `@teams` (formerly `@stats`) is
getting its value and it no longer is relying on any type of
pre-existing state.