Refactor #count to #update_records
This method does exactly the same thing but it’s no longer relying on
any outside data. This is the key point I’m trying to make with these
refactoring. You want to try to have your  methods rely on as little
outside state as possible. Pass in any of the information that your
method requires and it will be much easier to debug issues when they
arise and to change your code in the future.